backend/requirements.txt
# ==============================================================================
# AGENTIC TUTOR - Python Dependencies
# ==============================================================================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.0
aiomysql>=0.2.0
cryptography>=41.0.0

# Security & Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# AI/LLM
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.20
langsmith>=0.1.0
openai>=1.0.0

# Vector Database
chromadb>=0.4.0
sentence-transformers>=2.0.0

# File Processing
pypdf>=3.0.0
python-pptx>=0.6.0
python-docx>=1.0.0

# Video/Audio Transcription (GPU-optimized) - optional
# faster-whisper>=0.10.0
# torch>=2.0.0
# torchvision>=0.15.0

# Utilities
httpx>=0.25.0
pyyaml>=6.0
aiofiles>=23.0.0
pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
